requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
filelock==3.18.0
frozenlist==1.6.0
google-auth==2.40.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
marshmallow==3.26.1
mongoengine==0.29.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.78.1
orjson==3.10.18
packaging==24.2
pipenv==2025.0.2
platformdirs==4.3.8
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
yarl==1.20.0
zstandard==0.23.0
